zephyr: prj: increase size of mcuboot partition
After updating from v1.7.99-ncs4 to v1.9.99-ncs2 it seems that the
bootloader has increased in size.
Attempting to link would state that the FLASH area was 700 bytes too
small, so this commit increases the mcuboot partition size with 700
bytes, resolving the issue.

The details of the size growth and the consequences of this commit, I do
not know. It may be worthwhile investigating if we can decrease the
image size and drop this commit.