pm: add inequally sized primary and secondary partitions
The default configuration of MCUboot is intended to swap between two
application images. Hence it makes sense that the sizes of both the
primary and secondary partition are equal.

In the B&O use case, however, the secondary partition is only used to
store the net-core image, while the primary partition is used to store
the app-core image.
In this case it makes sense to allow inequal sizes as the net-core only
has 256kB flash while the default equally sized partitions typically
left around 450kB for the secondary image as well.

This is implemented as two Kconfigs. One to allow inequal sizes and one
to specify the size of the secondary partition. The primary partition
will eat up all residual space.