boot: zephyr: sysflash: restore FLASH_AREA_IMAGE_x macros
When introducing the MCUBOOT_MCUMGR_SIMPLE_IMAGE_INDEX option[1], the
following macros were changed:

        FLASH_AREA_IMAGE_PRIMARY(x)

        FLASH_AREA_IMAGE_SECONDARY(x)

These macros are used in other parts of the MCUboot codebase as well,
so it seems more appropriate to make the config implementation a bit
uglier to restore the intended behavior[2] of these macros.
The SIMPLE_IMAGE_INDEX is only relevant for `boot_serial.c` to simplify
the `mcumgr` image numbering.

Note: one case is unreachable and commented out since the SCRATCH
      partition definition also is not available here.

[1] https://github.com/AudioStreamingPlatform/sdk-mcuboot/commit/766be5d72bbb722aea365036588b5027ff1e5400
[2] https://github.com/nrfconnect/sdk-mcuboot/commit/7c3d7ed725919b5c0cce8688d565dec59957cb9b