boot: serial: write MCUboot trailer only once
The code that adds the MCUboot image trailer was run on _every chunk
recieved_ during an upload. This is unnecessary and wasteful in terms
of writing and erasing flash.

Let's move it to only run when the last chunk has been recieved.
It also only really makes sense to write the image trailer once the
complete image upload is done.