boot: serial: only add MCUboot trailer for NetCore-image
In the B&O setup, we are using all of the flash partitions in a bit of a
different manner than regular MCUboot examples[1].
Amongst other things, we add a custom B&O image trailer to the s0 and s1
partitions for the NRF immutable bootloader, b0, to read and react on.
Also, we have dropped the concept of a "secondary" slot which would
normally be used to flash both AppCore and NetCore image to, which
MCUboot would then perform image swapping on during the next boot. We
deploy directly to fixed partitions and don't utilise the image swapping
except for the (different) case of updating the NetCore image.

This conflicts with the MCUboot "standard" image trailer[2], so let's
only write the MCUboot trailer for the one image, where we use it;
NetCore.
The MCUboot trailer includes flags to signal wether a "swap", i.e.
update, should happen or not. So this trailer is what makes MCUboot sync
any updated NetCore image into the actual NetCore flash (performed in
combination with the b0n, NetCore bootloader).

[1] https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/mcuboot/design.html#multiple-image-boot
[2] https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/mcuboot/design.html#image-trailer