boot: bootutil: align validation of NetCore image update
In the B&O use case, we are using the flash partitions in an
unconventional way (as compared to "regular" MCUboot examples[1]).
This means that the assumption that there is a "secondary" slot present,
no longer holds. So we have to adapt the validation check to accept
correctly understand whether it is dealing with a NetCore update or not.

In our use case, any image located in the `mcuboot_secondary` partition
is _always_ an image for the NetCore, as we do direct uploads to either
the AppCore (`mcuboot_primary`) and NetCore (`mcuboot_secondary`).

[1] https://developer.nordicsemi.com/nRF_Connect_SDK/doc/latest/mcuboot/design.html#multiple-image-boot